Add AllRunningQueries component
Which is the same as MyOwnRuns, except it displays user
