+++ src/main/resources/assets/javascripts/components/AllRunningQueries.react.js
@@ -1,0 +1,15 @@
+/** @jsx React.DOM */
+var React = require('react');
+
+/* Components */
+var RunsTable = require('./RunsTable.react');
+
+var AllRunningQueries = React.createClass({
+  displayName: 'AllRunningQueries',
+
+  render() {
+    return <RunsTable forCurrentUser={false} />;
+  },
+});
+
+module.exports = AllRunningQueries;
